requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
boto3==1.34.154
botocore==1.34.154
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.6
django-cors-headers==4.3.1
django-debug-toolbar==4.4.2
django-storages==1.14.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Faker==26.0.0
faker-commerce==1.0.4
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.3.7
MarkupSafe==2.1.5
mysqlclient==2.2.4
pillow==10.4.0
prompt_toolkit==3.0.47
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
